+++ src/deployments/helm_deployment/nimlibp2p/experiments/regression/regression.py
@@ -19,17 +19,14 @@ from pydantic import BaseModel, ConfigDict, Field, PositiveInt
 from ruamel import yaml
 
 # Project Imports
-from src.deployments.core.kube_utils import (
-    dict_set,
-    get_cleanup,
-    get_future_time,
-    kubectl_apply,
-    wait_for_rollout,
-    wait_for_time,
-)
+from src.deployments.core.k8s_cleanup import get_cleanup
+from src.deployments.core.k8s_deploy import kubectl_apply
+from src.deployments.core.k8s_rollout import wait_for_rollout
 from src.deployments.experiments.base_experiment import BaseExperiment
 from src.deployments.helm_deployment.builders import build_deployment
 from src.deployments.registry import experiment
+from src.utils.dict_utils import dict_set
+from src.utils.time_utils import get_future_time, wait_for_time
 
 logger = logging.getLogger(__name__)
